logic werkte niet dus is weg
